Added Postgres Device and Account Managers (#149)

+++ server/src/config.rs
@@ -16,7 +16,7 @@ pub struct ServerConfig<T: StateType> {
 pub struct ServerCliConfig {
     pub address: Option<String>,
     pub link_secret: Option<String>,
-    pub provision_timeout: Option<u64>,
+    pub provision_timeout: Option<u32>,
     pub message_buffer_size: Option<usize>,
     pub tls: Option<TlsConfig>,
     pub logging: Option<String>,
@@ -26,7 +26,7 @@ impl ServerCliConfig {
     pub fn new(
         address: Option<String>,
         link_secret: Option<String>,
-        provision_timeout: Option<u64>,
+        provision_timeout: Option<u32>,
         message_buffer_size: Option<usize>,
         tls: Option<TlsConfig>,
         logging: Option<String>,
